requirements.txt
click>=8.1.7
pydantic>=2.10.0
pydantic-settings>=2.5.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
httpx>=0.27.0
sqlalchemy>=2.0.29
alembic>=1.13.1
aiosqlite>=0.20.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
prometheus-client>=0.20.0
bcrypt>=4.2.0
PyJWT>=2.8.0
stripe>=8.0.0
pydantic[email]>=2.10.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
ruff>=0.4.4
